`register_transformation`: add opt `~instrument`
Before, the way to register an instrumentation PPX was to pass the
implementation file transformation to the `~impl`-argument of
`Driver.register_transformation`. That way, the position at which it was
applied was determined as follows: a deriver registered through `~impl`
is a whole file transformation. All whole file transformations get
applied after `~rules` rewriters get applied. Among whole file
transformations, the order is alphabetic with respect to the name they
are registered with. So the only way to influence when an instrumentation
PPX got applied was through its name.

This commit adds an optional `~instrument`-argument to
`Driver.register_transformation`, that allows to specify if one wants the
instrumentation PPX to be applied before or after all other whole file
transfomrations and `rules` rewriters get applied.

Signed-off-by: Sonja Heinze <sonjaleaheinze@gmail.com>