Support binary AST's of any supported version
Before, the OCaml version of any binary ast input had to coincide with
the version the ppxlib driver got compiled with. That was a problem for
bucklescript/rescript.

With this commit, all versions from the Versions module are accepted as
input. The input version is preserved by ppxlib when outputting a binary
AST.

Signed-off-by: Sonja Heinze <sonjaleaheinze@gmail.com>

More idiomatic version of create_versioned_error_f

Signed-off-by: Jeremie Dimino <jeremie@dimino.org>

tweak

Signed-off-by: Jeremie Dimino <jeremie@dimino.org>