Fix cli cookies handling
Before, cookies handed in through the cli were added to the `Ast_mapper`
cookies reference right away when reading the cli arguments. But in the
case of a binary AST input, that reference gets reset to only contain
the cookies from the ppx context when parsing that context. Therefore,
cli cookies were ignored for binary AST inputs.

This commit fixes that by adding the cli cookies to the `Ast_mapper`
cookies reference after parsing the ppx context.

Signed-off-by: Sonja Heinze <sonjaleaheinze@gmail.com>