Fix parse error location
When there is a source parse error, the file name responsible for the error
is added to the error location. That's necessary to report the file name
even if the error is due to the file not existing. Before, that was done
by overwriting the error location by an empty ghost range located in the
given file and therefore messing up the start and end positions. This
commit fixes that.

Signed-off-by: Sonja Heinze <sonjaleaheinze@gmail.com>