Add test for 501 migrations
Currently, the test checks whether parsing on 5.1.0 directly is the same
as parsing, down-migrating and then up-migrating. That's why now the
test is only enabled on 5.1.0.

When bumbing ppxlib's AST to 5.1.0, we should enable it only for 5.0.0.
Then, the test will check whether parsing on 5.0.0 is the same as
parsing, up-migrating and then down-migrating.

Signed-off-by: Sonja Heinze <sonjaleaheinze@gmail.com>