Fix migrations between 5.0.0 and 5.1.0
The commit adapts to the parsetree modification in value binding constraints
that happened from [ocaml.5.1.0~alpha1] to [ocaml.5.1.0~alpha2].

It also fixes several cases, in which one of the two migrations were
wrong as well as the locations returned by those migrations.

It also adds tests to test both round-trips as well as a test comparing
parsing on 5.0.0 with parsing on 5.1.0 and then down migrating. This last
test will need to be updated when bumping the AST.

Signed-off-by: Sonja Heinze <sonjaleaheinze@gmail.com>

Co-authored-by: Florian Angeletti <florian.angeletti@inria.fr>
Co-authored-by: Paul-Elliot <peada@free.fr>