Add support for OCaml 5.1
* Update trunk AST for pvb_constraint
* Add migration from 501 to 500
* Bump the 501 AST magic numbers
* Binary AST reader: use latest possible AST version

When the magic numbers aren't unique, before this commit, the binary
AST reader would read into the first AST version matching the magic
number. With this commit, it takes the last such version. That's important
for trunk support.

* Fix inline tests' dependency on OCaml version

Done by making the expected output also dependent on the OCaml version

* Add a add tests to test both round-trips as well as a test comparing
parsing on 5.0.0 with parsing on 5.1.0 and then down migrating. This last
test will need to be updated when bumping the AST.

Co-authored-by: Samuel Hym <samuel.hym@rustyne.lautre.net>
Co-authored-by: Florian Angeletti <florian.angeletti@inria.fr>
Co-authored-by: Sonja Heinze <sonjaleaheinze@gmail.com>
Co-authored-by: Paul-Elliot <peada@free.fr>
Signed-off-by: Thibaut Mattio <thibaut.mattio@gmail.com>