Make Context_free.map_top_down embed_errors arg optional
This extra argument that was added in #453 caused breakage
in `ocsigen-i18n` that was using `Context_free.map_top_down`.
Although it's undocumented, this function is still part of the public
API. Adding an extra optional argument is not strictly speaking
non-breaking but is good enough of an effort considering the status
of this function and in particular, the fact that it only has a single
public external user for which it prevents breakage.

Signed-off-by: Nathan Rebours <nathan.p.rebours@gmail.com>