Make -unused-code-warning tests actual tests
In their previous form they were not executed as part of the default
build or as part of the runtest alias.

They relied on a deriving_inline workflow which wasn't great for tests
and it lacked a bit of context on what was being tested so I converted
them to cram tests. It's a bit harder to read the generated code than
it was but the whole setup is not spread across 3 different directories
anymore and the tests are now documented.

Signed-off-by: Nathan Rebours <nathan.p.rebours@gmail.com>