Context_free: add a special_function' variant
Parsing `Longindent.t` is quite unpractical because Longident.t
are used as an IR for different and distinct syntaxic classes. This
means that interpreting the string argument of the
`Context_free.special_function` is less trivial than it appears.

To circumvent this issue, this commit takes the view that in complex
cases the user themself can still easily specify the identifier that
they intended to use.

Thus it sounds sensible to provides a `special_function'` variant
function which takes directly a `Longident.t` argument for advanced
uses.

Signed-off-by: octachron <octa@polychoron.fr>