Change Location.none
Location.none has been changed to match the compiler's Location.none as
of OCaml 4.08. This has been stable for some time. We also provide a
Location.is_none which checks if a location is equal to Location.none
for all versions of the compiler we currently support (i.e. >= 4.04).

The decision to change Location.none has been made to enable the code to
be simplified if we decide to change the lower bound of ppxlib to 4.08
in the future.

Location.is_none has been used to fix loc_of_attribute which would
sometimes miss attributes with none locations (like compiler inserted
ocaml.doc attributes).

Signed-off-by: Patrick Ferris <patrick@sirref.org>