Add dev tool to generate dummy migration
This tool is inspired from the omp version of old.

It is invoked as follows:
```
dune exec -- dev/gencopy/gencopy.exe astlib/ast_X.ml astlib/ast_Y.ml > migrate_X_Y.ml
```
It will generates deep "copy" functions from ast version X to version Y.
These are dummy function that are generated under the assumption that
the ast types in X and Y are identical.

When adding support for new compiler versions, this should be used as a
base to be editted, following the compiler errors that will point to AST
nodes that actually differ between the two versions.

Signed-off-by: Nathan Rebours <nathan.p.rebours@gmail.com>