Add remove functions for migration metadata
Ppxlib stores migration metadata as attributes in the AST. Under certain
circumstances this metadata may still be in the output (if a migration
that introduces the attribute does not have a corresponding opposite
migration to remove it). Most of the time there is a corresponding
migration and/or the particular encoding disappears and we no longer
have the attribute at all.

Signed-off-by: Patrick Ferris <patrick@sirref.org>